backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
celery>=5.3
redis>=5.0
mysqlclient>=2.2
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
reportlab>=4.0
XlsxWriter>=3.1
gunicorn>=21.2
python-dotenv>=1.0
